// src/app/models/product-result.model.ts
export interface ProductResult {
    code: string
    errors: any[]
    product: Product
    result: Result
    status: string
    warnings: Warning[]
  }
// ...
  export interface Result {
    id: string
    lc_name: string
    name: string
  }
  
  export interface Warning {
    field: Field
    impact: Impact
    message: Message
  }
  
  export interface Field {
    id: string
    value: string
  }
  
  export interface Impact {
    id: string
    lc_name: string
    name: string
  }
  
  export interface Message {
    id: string
    lc_name: string
    name: string
  }